refactor(EventBadge): 매직 넘버들은 Contants 클래스 내에서 관리하도록 수정
