+++ src/main/java/christmas/Constants.java
@@ -1,0 +1,8 @@
+package christmas;
+
+public class Constants {
+    public static final int BADGE_LAST_CRITERA = 20000;
+    public static final int BADGE_MIDDEL_CRITERA = 10000;
+    public static final int BADGE_FIRST_CRITERA = 5000;
+
+}
